+++ packages/reactabular-utils/CHANGELOG.md
@@ -1,7 +1,7 @@
 7.1.0 / 2016-xx-xx
 ==================
 
-  * Feature - Add `columnsAreEqual` and `rowsAreEqual` checkers.
+  * Feature - Add `columnsAreEqual` checker.
   * Feature - `evaluateTransforms` will throw if all transforms aren't functions.
 
 7.0.0 / 2016-11-03