+++ packages/reactabular/CHANGELOG.md
@@ -32,7 +32,7 @@
 
 ## reactabular-utils
 
-  * Feature - Add `columnsAreEqual` and `rowsAreEqual` checkers.
+  * Feature - Add `columnsAreEqual` checker.
   * Feature - `evaluateTransforms` will throw if all transforms aren't functions.
 
 ## reactabular-resolve